code to to get notes and make song
Reads in screen data and gets the note that was played on the key board, and saves it to a file. The file can then be used to create a song for the red bot, it is not done automatically which might be something to add in futer. Was not need for class just thought it would be cool to do any song. It is recognizable but not the best quality, can not play 'cords'(multiple notes at same time) since my red bot only has one buzzer.